http: set flush_on_worker_key_change only when templateURL+batches are used
There are use-cases where the user wants to maintain ordering for certain
categories of messages. In that case, closing the batch early is not
necessary.

flush_on_worker_key_change can be changed from the HTTP destination's
config, we should only force-set it to TRUE when templated URLs are used
together with batches and multiple workers.

Signed-off-by: László Várady <laszlo.varady@anno.io>
Signed-off-by: Hofi <hofione@gmail.com>