light: add functional test for file logrotation
The test generates a message source that sends 1000 messages to the a
file destination with logrotate enabled. The logfile has a maximum file
size of 10KB and should be rotated 5 times.
After all messages have been send it is verified that all messages were
received and all created logfiles are smaller the defined maximum file
size.

Signed-off-by: Alexander Zikulnig <42011733+azike@users.noreply.github.com>