affile: add logrotate-function in affile-dest
React to log_pipe_notify with code 'NC_LOGROTATE' for AFFileDestWriter
objects.
Upon notification it is checked if logrotation should be performed based
on the settings stored in 'logrotate_options'. If a logrotation
operation is pending it is executed and the active logfile is openeded
again. The new fd is returned via the user_data arg pointer.

The function is intended to be called by the current logwriter worker thread
after finishing flushing data.
Note that affile_dw_reopen can not be called from within
log_writer_work_perform as it would result in a deadlock (worker waits
on itself to finish).

Signed-off-by: Alexander Zikulnig <42011733+azike@users.noreply.github.com>