cloud-auth: do no directly include picojson/picojson.h
There is a macro that modifies the number
parsing behavior of picojson (PICOJSON_USE_INT64).

When jwt-cpp/jwt.h includes picojson.h it explicitly
enables PICOJSON_USE_INT64. I did not check the code
thoroughly, but it is possible that they explicitly
build on some int64_t related functionality.

However if we directly include jwt.h without setting
PICOJSON_USE_INT64, picojson will not have int64_t
support. This resulted in a strange behavior of
having int64_t type hinted value in a json object,
but the type checker not recognizing it as a number.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>
Signed-off-by: Hofi <hofione@gmail.com>