poll_events: added a version that does not do any of the poll event jobs if its new notified member set to TRUE
This can be used whereas a PollEvents object is required and cannot be NULL, but its functionality is not needed at all, e.g in the file_reader, where the watched file modification events can be detected and delivered in other ways.
Far not the best solution, but PollEvents is too deeply integrated currently, and I did not want to do an even bigger refactor.

Signed-off-by: Hofi <hofione@gmail.com>