style-check: Fix locale-dependent behavior and portability
The script check the output of astlye(1) for messages in English, but
when the user locale is different, it will not successfully detect files
with style issues.  Ensure we have the default C locale before checking
style.

While here, make the script more portable by not requiring bash (which
might not be available as /bin/bash).  Rely on plain sh (which is known
to be at /bin/sh) and remove bashisms.

Signed-off-by: Romain Tartière <romain@blogreen.org>