affile: add logrotate types and rotation logic
Logrotation is performed in the function do_logrotate and for now
implements a simple size based rotation.
There is no in-memory storage or list to keep track of rotated files.
With each rotation the number of log files is established through the
filenames and their rotation suffixes.

The log files are first scanned starting with the main log file until
the first roation is found that is not yet 'used'.
Then the files are rotated up to that index. If there are 'holes' in
the sequence of log files (e.g. file.log, file.log.1, file.log.3 -->
file.log.2 is missing) the files that come after that missing one are
not rotated.
This should prevent the deletion of logfiles in case a reopen attempt
fails.

Signed-off-by: Alexander Zikulnig <42011733+azike@users.noreply.github.com>