affile: add logrotate function in affile-dest
React to log_pipe_notify with code NC_LOGROTATE for AFFileDestWriter
objects.
Upon notification it is checked if logrotation should be performed based
on the settings stored in logrotate_options and the current cached
filesize. The function is intended to be called by the current logwriter
worker thread when flushing log messages.

If logrotation and reopen of the log file were successful the caller is
informed trough log_pipe_notify return codes.

Note that affile_dw_reopen can not be called from within
log_writer_work_perform as it would result in a deadlock (worker waits
on itself to finish).

Signed-off-by: Alexander Zikulnig <42011733+azike@users.noreply.github.com>